requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
slowapi
httpx
python-dotenv
google-genai
duckduckgo-search
pydantic